+++ src/lib.rs
@@ -51,9 +51,9 @@ SOFTWARE.
 use lazy_static::lazy_static;
 use std::collections::HashSet;
 use std::future::Future;
-use std::sync::atomic::AtomicU64;
+use std::sync::atomic::{AtomicU64, Ordering};
 use std::sync::Mutex;
-use tokio::time::Duration;
+use tokio::time::{Duration, Interval};
 
 /// **INTERNAL** Use macro [`set_timeout`] instead!
 ///
@@ -90,10 +90,10 @@ lazy_static! {
     };
 }
 
-/// Creates a future that glues the tokio interval function together with
-/// the provided callback. Helper function for [`_set_interval_spawn`].
-async fn _set_interval(f: impl Fn() + Send + 'static, ms: u64, id: u64) {
-    // own scope -> early drop lock
+/// Common code for [_set_interval] and [_set_interval_async]. Adds the new interval to the
+/// interval manager.
+async fn _set_interval_common(ms: u64, id: u64) -> Interval {
+    // dedicated scope -> early drop lock
     {
         let mut map = INTERVAL_MANAGER.running_intervals.lock().unwrap();
         map.insert(id);
@@ -105,13 +105,21 @@ async fn _set_interval(f: impl Fn() + Send + 'static, ms: u64, id: u64) {
     // similar to Javascript in this library, we work around this.
     int.tick().await;
 
+    int
+}
+
+/// Creates a future that glues the tokio interval function together with
+/// the provided callback. Helper function for [`_set_interval_spawn`].
+async fn _set_interval(f: impl Fn() + Send + 'static, ms: u64, id: u64) {
+    let mut int = _set_interval_common(ms, id).await;
+
     // this looks like it runs for ever, but this is not how tokio works
     // at least with tokio 1.6.0 and Rust 1.52 this stops executing
     // when the tokio runtime gets dropped.
     loop {
         int.tick().await;
 
-        // - breaks the loop if the interval war cleared.
+        // - breaks the loop if the interval was cleared.
         // - dedicated block to drop lock early
         {
             let map = INTERVAL_MANAGER.running_intervals.lock().unwrap();
@@ -124,15 +132,55 @@ async fn _set_interval(f: impl Fn() + Send + 'static, ms: u64, id: u64) {
     }
 }
 
+/// Async version of [_set_interval].
+async fn _set_interval_async<T, Func, Fut>(f: Func, ms: u64, id: u64)
+where
+    Func: (Fn() -> Fut) + Send + 'static + Sync,
+    Fut: Future<Output = T> + Send + Sync,
+{
+    let mut int = _set_interval_common(ms, id).await;
+
+    // this looks like it runs for ever, but this is not how tokio works
+    // at least with tokio 1.6.0 and Rust 1.52 this stops executing
+    // when the tokio runtime gets dropped.
+    loop {
+        int.tick().await;
+
+        // - breaks the loop if the interval was cleared.
+        // - dedicated block to drop lock early
+        {
+            let map = INTERVAL_MANAGER.running_intervals.lock().unwrap();
+            if !map.contains(&id) {
+                break;
+            }
+        }
+
+        f().await;
+    }
+}
+
 /// **INTERNAL** Use macro [`set_interval`] instead!
 ///
 /// Acquires a new ID, creates an interval future and returns the ID. The future gets
 /// dispatched as tokio task.
 pub fn _set_interval_spawn(f: impl Fn() + Send + 'static, ms: u64) -> u64 {
-    let id = INTERVAL_MANAGER
-        .counter
-        .fetch_add(1, std::sync::atomic::Ordering::SeqCst);
+    // assign next ID to interval
+    let id = INTERVAL_MANAGER.counter.fetch_add(1, Ordering::SeqCst);
     tokio::spawn(_set_interval(f, ms, id));
+    id
+}
+
+/// **INTERNAL** Use macro [`set_interval_async`] instead!
+///
+/// Async version of [_set_interval_spawn].
+pub fn _set_interval_spawn_async<T: 'static, Func, Fut>(f: Func, ms: u64) -> u64
+where
+    Func: (Fn() -> Fut) + Send + 'static + Sync,
+    Fut: Future<Output = T> + Send + 'static + Sync,
+{
+    // assign next ID to interval
+    let id = INTERVAL_MANAGER.counter.fetch_add(1, Ordering::SeqCst);
+    tokio::spawn(_set_interval_async(f, ms, id));
     id
 }
 
@@ -237,15 +285,17 @@ macro_rules! set_timeout_async {
        };
 
 
-    /* There are no async expressions, I think
-    // match for expr, like `set_timeout!(println!())`
-    ($cb:expr, $ms:literal) => {
-        $crate::set_timeout_async!(|| $cb, $ms);
-    };*/
-    // match for async block
-    (async $cb:block, $ms:literal) => {
-        $crate::set_timeout_async!(async || $cb, $ms);
-    };
+       /* There are no async expressions, I think
+       // match for expr, like `set_timeout!(println!())`
+       ($cb:expr, $ms:literal) => {
+           $crate::set_timeout_async!(|| $cb, $ms);
+       };*/
+
+       // match for async block
+       (async $cb:block, $ms:literal) => {
+           $crate::set_timeout_async!(async || $cb, $ms);
+       };
+       */
 }
 
 /// Creates a timeout that behaves similar to `setInterval(callback, ms)` in Javascript
@@ -271,7 +321,7 @@ macro_rules! set_timeout_async {
 ///
 /// #[tokio::main]
 /// async fn main() {
-///     set_interval!(println!("hello1"), 50);
+///     let interval = set_interval!(println!("hello1"), 50);
 ///     // If you want to clear the interval later: save the ID
 ///     // let id = set_interval!(println!("hello1"), 50);
 ///     println!("hello2");
@@ -297,17 +347,60 @@ macro_rules! set_interval {
     };
     // match for expr, like `set_interval!(println!())`
     ($cb:expr, $ms:literal) => {
-        $crate::set_interval!(|| $cb, $ms);
+        $crate::set_interval!(|| $cb, $ms)
     };
     // match for block
     ($cb:block, $ms:literal) => {
-        $crate::set_interval!(|| $cb, $ms);
+        $crate::set_interval!(|| $cb, $ms)
     };
+}
+
+/// Async version of [set_interval].
+///
+/// # Example
+/// ```rust
+/// use tokio::time::Duration;
+/// use tokio_js_set_interval::set_interval_async;
+///
+/// #[tokio::main]
+/// async fn main() {
+///     async fn async_foo() {
+///         println!("hello1")
+///     }
+///
+///     set_interval_async!(async_foo, 50);
+///     // If you want to clear the interval later: save the ID
+///     // let id = set_interval_async!(async_foo, 50);
+///     println!("hello2");
+///     // prevent that tokios runtime gets dropped too early
+///     // "hello1" should get printed 2 times (50*2 == 100 < 120)
+///     tokio::time::sleep(Duration::from_millis(120)).await;
+/// }
+/// ```
+#[macro_export]
+macro_rules! set_interval_async {
+    // match for identifier, i.e. a closure, that is behind a variable
+    ($cb:ident, $ms:literal) => {
+        $crate::_set_interval_spawn_async($cb, $ms)
+    }; /* Async closures not supported currently:
+          https://github.com/rust-lang/rust/issues/62290
+       // match for direct closure expression
+       (|| $cb:expr, $ms:literal) => {
+           $crate::_set_interval_spawn(|| $cb, $ms)
+       };
+       // match for direct move closure expression
+       (move || $cb:expr, $ms:literal) => {
+           $crate::_set_interval_spawn(move || $cb, $ms)
+       };
+       // match for block
+       ($cb:block, $ms:literal) => {
+           $crate::set_interval!(|| $cb, $ms)
+       };*/
 }
 
 #[cfg(test)]
 mod tests {
-    use std::sync::atomic::{AtomicU64, Ordering};
+    use std::sync::atomic::AtomicU64;
     use std::sync::Arc;
 
     use super::*;
@@ -356,6 +449,12 @@ mod tests {
         // macro takes identifiers (which must point to closures)
         let closure = || println!("hello5");
         set_interval!(closure, 1);
+
+        // macro takes identifiers (which must point to closures)
+        async fn async_foo() {
+            println!("hello5");
+        }
+        set_interval_async!(async_foo, 1);
     }
 
     /// Test can't been automated because the test is correct
@@ -459,6 +558,22 @@ mod tests {
         // give the tokio runtime enough execution time to execute the interval 3 times
         tokio::time::sleep(Duration::from_millis(180)).await;
         assert_eq!(counter.load(Ordering::SeqCst), 3);
+    }
+
+    #[tokio::test]
+    async fn test_set_interval_async() {
+        static COUNTER: AtomicU64 = AtomicU64::new(0);
+        async fn async_foo() {
+            COUNTER.fetch_add(1, Ordering::SeqCst);
+        }
+        {
+            // the block is required to change the return type to "()"
+            set_interval_async!(async_foo, 50);
+        }
+
+        // give the tokio runtime enough execution time to execute the interval 3 times
+        tokio::time::sleep(Duration::from_millis(180)).await;
+        assert_eq!(COUNTER.load(Ordering::SeqCst), 3);
     }
 
     #[tokio::test]